Integrate test file utils into main set
