+++ collect_app/src/androidTest/java/org/odk/collect/android/feature/formentry/IntentGroupTest.java
@@ -37,7 +37,7 @@ import static org.hamcrest.core.StringEndsWith.endsWith;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 import static org.odk.collect.android.support.matchers.CustomMatchers.withIndex;
-import static org.odk.collect.android.support.FileUtils.copyFileFromAssets;
+import static org.odk.collect.android.utilities.FileUtils.copyFileFromAssets;
 
 import android.app.Activity;
 import android.app.Instrumentation;
@@ -50,6 +50,7 @@ import android.os.Environment;
 import androidx.core.content.FileProvider;
 import androidx.test.core.app.ApplicationProvider;
 import androidx.test.espresso.matcher.ViewMatchers;
+import androidx.test.platform.app.InstrumentationRegistry;
 
 import org.junit.Rule;
 import org.junit.Test;
@@ -251,7 +252,7 @@ public class IntentGroupTest {
                 .getExternalFilesDir(Environment.DIRECTORY_DOWNLOADS);
 
         File file = File.createTempFile(name, extension, downloadsDir);
-        copyFileFromAssets("media" + File.separator + name + "." + extension, file.getPath());
+        copyFileFromAssets(InstrumentationRegistry.getInstrumentation().getContext(), file.getPath(), "media" + File.separator + name + "." + extension);
         return getUriForFile(file);
     }
