+++ collect_app/src/androidTest/java/org/odk/collect/android/support/StorageUtils.kt
@@ -17,11 +17,13 @@ package org.odk.collect.android.support
 
 import android.app.Application
 import androidx.test.core.app.ApplicationProvider
+import androidx.test.platform.app.InstrumentationRegistry
 import org.apache.commons.csv.CSVFormat
 import org.apache.commons.csv.CSVRecord
 import org.odk.collect.android.formmanagement.LocalFormUseCases
 import org.odk.collect.android.injection.DaggerUtils
 import org.odk.collect.android.storage.StorageSubdirectory
+import org.odk.collect.android.utilities.FileUtils
 import java.io.File
 import java.io.FileReader
 import java.io.IOException
@@ -83,13 +85,20 @@ object StorageUtils {
     fun copyInstance(instanceFileName: String, projectName: String) {
         val instanceDirPath = getInstancesDirPath(projectName) + instanceFileName.split("\\.".toRegex()).toTypedArray()[0]
         File(instanceDirPath).mkdir()
-        FileUtils.copyFileFromAssets("instances/$instanceFileName", "$instanceDirPath/$instanceFileName")
+        FileUtils.copyFileFromAssets(
+            InstrumentationRegistry.getInstrumentation().getContext(),
+            "$instanceDirPath/$instanceFileName",
+            "instances/$instanceFileName"
+        )
     }
 
     @Throws(IOException::class)
     private fun copyForm(formFilename: String, copyTo: String, projectName: String): String {
         val pathname = getFormsDirPath(projectName) + copyTo
-        FileUtils.copyFileFromResources("forms/$formFilename", pathname)
+        FileUtils.copyFileFromResources(
+            "forms/$formFilename",
+            pathname
+        )
         return pathname
     }
 
@@ -98,7 +107,11 @@ object StorageUtils {
         val mediaPathName = getFormsDirPath(projectName) + formFilename.replace(".xml", "") + org.odk.collect.android.utilities.FileUtils.MEDIA_SUFFIX + "/"
         org.odk.collect.android.utilities.FileUtils.checkMediaPath(File(mediaPathName))
         for (mediaFilePath in mediaFilePaths) {
-            FileUtils.copyFileFromAssets("media/$mediaFilePath", mediaPathName + getMediaFileName(mediaFilePath))
+            FileUtils.copyFileFromAssets(
+                InstrumentationRegistry.getInstrumentation().getContext(),
+                mediaPathName + getMediaFileName(mediaFilePath),
+                "media/$mediaFilePath"
+            )
         }
     }
